refactor: MAT-269 선수 교체 Store를 features 폴더로 이관 (내부 구현), Policy 파일 이름 개선
