+++ src/features/playerSubstitution/index.ts
@@ -1,3 +1,2 @@
 export * from './components';
-export * from './PlayerSubstitutionPolicies';
-export * from './PlayerSubstitutionAdapter';
+export * from './PlayerSubstitutionAdapter'; // NOTE: 내부용으로만 사용
