+++ src/stores/index.ts
@@ -1,5 +1,4 @@
 export * from './selectedPlayer';
-export * from './playerSubstitution';
 export * from './pageTitle';
 export * from './loggedInUser';
 export * from './userSettingsModal';
